tiledb parsing / docs
